+++ src/lib/components/button/NewProposalButton.tsx
@@ -4,14 +4,14 @@ import { Tooltip } from "../Tooltip";
 import { useInternalNavigate } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
-import { useGovParams } from "lib/services/proposalService";
-import { AccessConfigPermission } from "lib/types";
+// import { useGovParams } from "lib/services/proposalService";
+// import { AccessConfigPermission } from "lib/types";
 
 export const NewProposalButton = () => {
   const navigate = useInternalNavigate();
-  const { data: govParams } = useGovParams();
-  const isPermissionless =
-    govParams?.uploadAccess.permission === AccessConfigPermission.EVERYBODY;
+  // const { data: govParams } = useGovParams();
+  // const isPermissionless =
+  //   govParams?.uploadAccess.permission === AccessConfigPermission.EVERYBODY;
 
   return (
     <Menu>
@@ -44,15 +44,16 @@ export const NewProposalButton = () => {
         >
           To Instantiate Contract
         </MenuItem> */}
-        <Tooltip
+        {/* <Tooltip
           label={
             isPermissionless
               ? "Not available in permissionless network"
               : undefined
           }
-        >
+        > */}
+        <Tooltip label="Coming soon!">
           <MenuItem
-            isDisabled={isPermissionless}
+            isDisabled
             icon={<CustomIcon name="admin" color="gray.600" />}
             onClick={() => {
               navigate({