+++ src/lib/pages/home/index.tsx
@@ -7,7 +7,6 @@ import { CURR_THEME } from "env";
 import { useCelatoneApp, useInternalNavigate } from "lib/app-provider";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
 import { CustomIcon } from "lib/components/icon";
-import { DummyJsonSchema } from "lib/components/json-schema";
 import PageContainer from "lib/components/PageContainer";
 import { ViewMore } from "lib/components/table";
 import { Tooltip } from "lib/components/Tooltip";
@@ -220,7 +219,6 @@ const Home = () => {
           />
         </Flex>
       </Flex>
-      <DummyJsonSchema />
       {isDevMode && (
         <section style={{ marginBottom: "48px" }}>
           <Flex gap={4} direction="column">